Eliminados atributos innecesarios en la interfaz IMessage y agregado IMessage::id() Eliminado el driver para TelApi

+++ DataServer/watcher/RemoteWatcher.hpp
@@ -17,8 +17,8 @@ namespace Gateway
 
     signals:
         void messageReceived(const IMessage *message);
-        void messageSent(uint messageId);
-        void messageRefused(uint messageId);
+        void messageSent(qlonglong messageId);
+        void messageRefused(qlonglong messageId);
     };
 
 }
